backend/requirements.txt
# Backend dependencies (pinned for reproducibility)
# Python: 3.12 compatible, FastAPI with Pydantic v1

fastapi==0.118.0
starlette==0.48.0
uvicorn[standard]==0.37.0
SQLAlchemy==2.0.41
pydantic==1.10.24
httpx==0.28.1
# Optional tooling used by the stack (already present in environment)
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0

